final_project/server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
waitress==3.0.0
openai==1.61.0